Use joblib.Parallel within Registry.compute
